Split api code from express example app
