+++ apps/airview-api-demo/README.md
@@ -1,0 +1,1 @@
+Example application to demonstrate the capabilities of the package 'airview-cms-api'
