+++ apps/airview-api-demo/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "airview-api-demo",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/AirWalk-Digital/airview.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/AirWalk-Digital/airview/issues"
+  },
+  "homepage": "https://github.com/AirWalk-Digital/airview#readme",
+  "dependencies": {
+    "airview-cms-api": "file:../../packages/airview-cms-api",
+    "dotenv": "^8.2.0",
+    "express": "^4.18.1"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.13",
+    "typescript": "^4.6.4"
+  }
+}
