+++ packages/airview-cms-api/package.json
@@ -2,10 +2,14 @@
   "name": "airview-cms-api",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "dist/index.js",
   "type": "module",
+  "files": [
+    "./dist"
+  ],
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "rollup -c"
   },
   "repository": {
     "type": "git",
@@ -19,15 +23,13 @@
   "homepage": "https://github.com/AirWalk-Digital/airview#readme",
   "dependencies": {
     "aws-sdk": "^2.1141.0",
-    "dotenv": "^8.2.0",
-    "express": "^4.18.1",
     "gray-matter": "^4.0.3",
     "jsonwebtoken": "^8.5.1",
     "node-fetch": "^3.2.4",
     "sqlite3": "^5.0.8"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
+    "@rollup/plugin-typescript": "^8.3.2",
     "@types/jsonwebtoken": "^8.5.1",
     "@types/node-fetch": "^2.6.1",
     "@types/sqlite3": "^3.1.8",
